🚚 Move components to correct folder
